Move MakeInternalFuncsStatic to separate file
