+++ src/opt.cpp
@@ -32,8 +32,7 @@
 */
 
 /** @file opt.cpp
-    @brief Implementations of various ispc optimization passes that operate
-           on the LLVM IR.
+    @brief Implementation of ISPC optimization pipeline.
 */
 
 #include "opt.h"
@@ -129,7 +128,6 @@ using namespace ispc;
 static llvm::Pass *CreateGatherCoalescePass();
 
 static llvm::Pass *CreateIsCompileTimeConstantPass(bool isLastTry);
-static llvm::Pass *CreateMakeInternalFuncsStaticPass();
 
 #ifndef ISPC_NO_DUMPS
 static llvm::Pass *CreateDebugPass(char *output);
@@ -1781,211 +1779,6 @@ static llvm::Pass *CreateDebugPassFile(int number, llvm::StringRef name, std::st
     return new DebugPassFile(number, name, dir);
 }
 #endif
-
-///////////////////////////////////////////////////////////////////////////
-// MakeInternalFuncsStaticPass
-
-/** There are a number of target-specific functions that we use during
-    these optimization passes.  By the time we are done with optimization,
-    any uses of these should be inlined and no calls to these functions
-    should remain.  This pass marks all of these functions as having
-    private linkage so that subsequent passes can eliminate them as dead
-    code, thus cleaning up the final code output by the compiler.  We can't
-    just declare these as static from the start, however, since then they
-    end up being eliminated as dead code during early optimization passes
-    even though we may need to generate calls to them during later
-    optimization passes.
- */
-class MakeInternalFuncsStaticPass : public llvm::ModulePass {
-  public:
-    static char ID;
-    MakeInternalFuncsStaticPass(bool last = false) : ModulePass(ID) {}
-
-    void getAnalysisUsage(llvm::AnalysisUsage &AU) const { AU.setPreservesCFG(); }
-
-    llvm::StringRef getPassName() const { return "Make internal funcs \"static\""; }
-    bool runOnModule(llvm::Module &m);
-};
-
-char MakeInternalFuncsStaticPass::ID = 0;
-
-bool MakeInternalFuncsStaticPass::runOnModule(llvm::Module &module) {
-    const char *names[] = {
-        "__avg_up_uint8",
-        "__avg_up_int8",
-        "__avg_up_uint16",
-        "__avg_up_int16",
-        "__avg_down_uint8",
-        "__avg_down_int8",
-        "__avg_down_uint16",
-        "__avg_down_int16",
-        "__fast_masked_vload",
-        "__gather_factored_base_offsets32_i8",
-        "__gather_factored_base_offsets32_i16",
-        "__gather_factored_base_offsets32_i32",
-        "__gather_factored_base_offsets32_i64",
-        "__gather_factored_base_offsets32_half",
-        "__gather_factored_base_offsets32_float",
-        "__gather_factored_base_offsets32_double",
-        "__gather_factored_base_offsets64_i8",
-        "__gather_factored_base_offsets64_i16",
-        "__gather_factored_base_offsets64_i32",
-        "__gather_factored_base_offsets64_i64",
-        "__gather_factored_base_offsets64_half",
-        "__gather_factored_base_offsets64_float",
-        "__gather_factored_base_offsets64_double",
-        "__gather_base_offsets32_i8",
-        "__gather_base_offsets32_i16",
-        "__gather_base_offsets32_i32",
-        "__gather_base_offsets32_i64",
-        "__gather_base_offsets32_half",
-        "__gather_base_offsets32_float",
-        "__gather_base_offsets32_double",
-        "__gather_base_offsets64_i8",
-        "__gather_base_offsets64_i16",
-        "__gather_base_offsets64_i32",
-        "__gather_base_offsets64_i64",
-        "__gather_base_offsets64_half",
-        "__gather_base_offsets64_float",
-        "__gather_base_offsets64_double",
-        "__gather32_i8",
-        "__gather32_i16",
-        "__gather32_i32",
-        "__gather32_i64",
-        "__gather32_half",
-        "__gather32_float",
-        "__gather32_double",
-        "__gather64_i8",
-        "__gather64_i16",
-        "__gather64_i32",
-        "__gather64_i64",
-        "__gather64_half",
-        "__gather64_float",
-        "__gather64_double",
-        "__gather_elt32_i8",
-        "__gather_elt32_i16",
-        "__gather_elt32_i32",
-        "__gather_elt32_i64",
-        "__gather_elt32_half",
-        "__gather_elt32_float",
-        "__gather_elt32_double",
-        "__gather_elt64_i8",
-        "__gather_elt64_i16",
-        "__gather_elt64_i32",
-        "__gather_elt64_i64",
-        "__gather_elt64_half",
-        "__gather_elt64_float",
-        "__gather_elt64_double",
-        "__masked_load_i8",
-        "__masked_load_i16",
-        "__masked_load_i32",
-        "__masked_load_i64",
-        "__masked_load_half",
-        "__masked_load_float",
-        "__masked_load_double",
-        "__masked_store_i8",
-        "__masked_store_i16",
-        "__masked_store_i32",
-        "__masked_store_i64",
-        "__masked_store_half",
-        "__masked_store_float",
-        "__masked_store_double",
-        "__masked_store_blend_i8",
-        "__masked_store_blend_i16",
-        "__masked_store_blend_i32",
-        "__masked_store_blend_i64",
-        "__masked_store_blend_half",
-        "__masked_store_blend_float",
-        "__masked_store_blend_double",
-        "__scatter_factored_base_offsets32_i8",
-        "__scatter_factored_base_offsets32_i16",
-        "__scatter_factored_base_offsets32_i32",
-        "__scatter_factored_base_offsets32_i64",
-        "__scatter_factored_base_offsets32_half",
-        "__scatter_factored_base_offsets32_float",
-        "__scatter_factored_base_offsets32_double",
-        "__scatter_factored_base_offsets64_i8",
-        "__scatter_factored_base_offsets64_i16",
-        "__scatter_factored_base_offsets64_i32",
-        "__scatter_factored_base_offsets64_i64",
-        "__scatter_factored_base_offsets64_half",
-        "__scatter_factored_base_offsets64_float",
-        "__scatter_factored_base_offsets64_double",
-        "__scatter_base_offsets32_i8",
-        "__scatter_base_offsets32_i16",
-        "__scatter_base_offsets32_i32",
-        "__scatter_base_offsets32_i64",
-        "__scatter_base_offsets32_half",
-        "__scatter_base_offsets32_float",
-        "__scatter_base_offsets32_double",
-        "__scatter_base_offsets64_i8",
-        "__scatter_base_offsets64_i16",
-        "__scatter_base_offsets64_i32",
-        "__scatter_base_offsets64_i64",
-        "__scatter_base_offsets64_half",
-        "__scatter_base_offsets64_float",
-        "__scatter_base_offsets64_double",
-        "__scatter_elt32_i8",
-        "__scatter_elt32_i16",
-        "__scatter_elt32_i32",
-        "__scatter_elt32_i64",
-        "__scatter_elt32_half",
-        "__scatter_elt32_float",
-        "__scatter_elt32_double",
-        "__scatter_elt64_i8",
-        "__scatter_elt64_i16",
-        "__scatter_elt64_i32",
-        "__scatter_elt64_i64",
-        "__scatter_elt64_half",
-        "__scatter_elt64_float",
-        "__scatter_elt64_double",
-        "__scatter32_i8",
-        "__scatter32_i16",
-        "__scatter32_i32",
-        "__scatter32_i64",
-        "__scatter32_half",
-        "__scatter32_float",
-        "__scatter32_double",
-        "__scatter64_i8",
-        "__scatter64_i16",
-        "__scatter64_i32",
-        "__scatter64_i64",
-        "__scatter64_half",
-        "__scatter64_float",
-        "__scatter64_double",
-        "__prefetch_read_varying_1",
-        "__prefetch_read_varying_2",
-        "__prefetch_read_varying_3",
-        "__prefetch_read_varying_nt",
-        "__prefetch_write_varying_1",
-        "__prefetch_write_varying_2",
-        "__prefetch_write_varying_3",
-        "__keep_funcs_live",
-#ifdef ISPC_XE_ENABLED
-        "__masked_load_blend_i8",
-        "__masked_load_blend_i16",
-        "__masked_load_blend_i32",
-        "__masked_load_blend_i64",
-        "__masked_load_blend_half",
-        "__masked_load_blend_float",
-        "__masked_load_blend_double",
-#endif
-    };
-
-    bool modifiedAny = false;
-    int count = sizeof(names) / sizeof(names[0]);
-    for (int i = 0; i < count; ++i) {
-        llvm::Function *f = m->module->getFunction(names[i]);
-        if (f != NULL && f->empty() == false) {
-            f->setLinkage(llvm::GlobalValue::InternalLinkage);
-            modifiedAny = true;
-        }
-    }
-
-    return modifiedAny;
-}
-
-static llvm::Pass *CreateMakeInternalFuncsStaticPass() { return new MakeInternalFuncsStaticPass; }
 
 /** Given an llvm::Value known to be an integer, return its value as
     an int64_t.
