+++ src/opt/MakeInternalFuncsStatic.h
@@ -31,15 +31,31 @@
    SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-/** @file ISPCPasses.h
-    @brief Includes available ISPC passes
-*/
-
 #pragma once
 
-#include "ImproveMemoryOps.h"
-#include "InstructionSimplify.h"
-#include "IntrinsicsOpt.h"
-#include "MakeInternalFuncsStatic.h"
-#include "Peephole.h"
-#include "ReplacePseudoMemoryOps.h"
+#include "ISPCPass.h"
+
+namespace ispc {
+/** There are a number of target-specific functions that we use during
+    these optimization passes.  By the time we are done with optimization,
+    any uses of these should be inlined and no calls to these functions
+    should remain.  This pass marks all of these functions as having
+    private linkage so that subsequent passes can eliminate them as dead
+    code, thus cleaning up the final code output by the compiler.  We can't
+    just declare these as static from the start, however, since then they
+    end up being eliminated as dead code during early optimization passes
+    even though we may need to generate calls to them during later
+    optimization passes.
+ */
+class MakeInternalFuncsStaticPass : public llvm::ModulePass {
+  public:
+    static char ID;
+    MakeInternalFuncsStaticPass(bool last = false) : ModulePass(ID) {}
+
+    void getAnalysisUsage(llvm::AnalysisUsage &AU) const { AU.setPreservesCFG(); }
+
+    llvm::StringRef getPassName() const { return "Make internal funcs \"static\""; }
+    bool runOnModule(llvm::Module &m);
+};
+llvm::Pass *CreateMakeInternalFuncsStaticPass();
+} // namespace ispc
